fix: cast dongle cd to long long before multiplication to avoid overflow

+++ src/routine_cdr.c
@@ -10,8 +10,8 @@ static void	release_dongles(t_coder *coder)
 	codexion = coder->codexion;
 	pthread_mutex_lock(&coder->dongle_pair.first->when_available_lock);
 	pthread_mutex_lock(&coder->dongle_pair.second->when_available_lock);
-	coder->dongle_pair.first->when_available = ft_get_time() + codexion->args.dongle_cooldown * 1000;
-	coder->dongle_pair.second->when_available = ft_get_time() + codexion->args.dongle_cooldown * 1000;
+	coder->dongle_pair.first->when_available = ft_get_time() + (long long)codexion->args.dongle_cooldown * 1000;
+	coder->dongle_pair.second->when_available = ft_get_time() + (long long)codexion->args.dongle_cooldown * 1000;
 	pthread_mutex_unlock(&coder->dongle_pair.second->when_available_lock);
 	pthread_mutex_unlock(&coder->dongle_pair.first->when_available_lock);
 	pthread_mutex_unlock(&coder->dongle_pair.second->owner_lock);
